Repairing infix "in" operator
The list of generators sent to the constructor could sometimes be vectors of FFEs rather than codewords.  This lead to a subtle error where a
Code might not be seen as having the right Family to be a container for Codewords, thus a difficult to debug error cropped up where "c in C" returned false even though c was definitely in C!  See the discussion in Issue #5 BCH decoding.