Fix PutStandardForm *again*
Commit 8474420b5faf7469e841e6e3674bed0e050db0c6 was meant to "fix"
`PutStandardForm` according to the commit message, although it is
not specified what it was supposed to fix.

This led to changed behavior of `PutStandardForm`. This was meant
to be fixed by c2b498f3d1635410c48461e3c8300752764b8c7d but it was
incorrect, too, now immediately leading to errors (it tried to use
the return value of `TriangulizeMat`, but this procedure never returns
anything)

Let's hope that three is the charm...