Fix out of bounds access in the C code
A compiler warning flagged this one: the global array relatorSelection has
length 5, but the code tried to access `relatorSelection[5]` which would be
the 6th element, as C uses 0-based indexing.

Upon inspecting the code, it also became clear that `relatorSelection[3]`
never is accessed, and that access to `relatorSelection[4]` and
`relatorSelection[5]` only occur together, in two places, and both are
used to modify `selectionPriority`.

Based on this, I believe both of those were off-by-one, and adjusted the code
accordingly