desauto: initialize the comment array to a sensible value
If left uninitialized, the comment array is passed down to client code
without being even written to, which means it contains whatever happens
to be on the stack whenever the frame for `main` is created. Recent
changes to glibc start code have changed said content, and the new one
happens to break the test suite when propagated into some temporary
files that are then fed to the main leonconv binary.

That explains why using ThreadSanitizer "fixes" the test suite:
presumably, the runtime initialization code happens to write in that
aread of the stack and changes yet again the contents of the comment
array, which happens to suit whatever expectations are in leonconv.

Fixes #98