Fixes for FerreroDesignCode() and QQRCode()
Ferrero design stuff was commented out ~2006 because it had a dependency
on the Sonata package.  It now works fine since Sonata is autoloaded by
Guava.  The QQRCode tests failed because the constructor uses
BlockMatrix() to build the generator matrix and the returned matrix
doesn't support Display().  Used MatrixByBlockMatrix() to convert to a
more friendly form.  Also reverted changes to LinearCodeByGenerators(),
the calling functions must handle setting the generator matrix...