Fix PutStandardForm(mat)  -- hopefully for the last time
This function has been the source of many bugs over the years.
There have been several iterations of one dev changing it and another
reverting those changes.  The argument to the function must be mutable
because the function needs to change it - the return value is the
column permutation that is needed during the Gaussian elimination
process to achieve standard form.  Added an Error() message so that
when calling code provides an immutable input we can track down
the actual source of the problem.